Import historical inventory counts, add an Orders tab for sellable items
Widen inventory_count_items.counted from integer to NUMERIC(14,8) - the
historical count export has fractional (weight-based) and even negative
values, and retaining that precision was the point of doing this import
at all. Trim the trailing zeros a fixed-scale column always renders with
on both places counted is displayed (Overview's "Current on hand" and
the Counts tab).

Add scripts/import_inventory_counts_csv.py for the raw count-log export
(SKU, Timestamp, Email, Amount, Notes - one row per item per submission,
no explicit session column). Rows sharing an exact timestamp become one
InventoryCount (verified real submitters never share a timestamp across
different people, so this is a safe grouping); each row becomes one
InventoryCountItem under it, backdated to that timestamp. Email has no
column of its own on either table, so it's folded into notes as a
"[email]" prefix ahead of the row's own notes rather than dropped -
that's the "clean up later" part. A SKU with no matching item gets a
stub Item created (sku only) so no count data is lost. Safe to re-run:
reuses an existing count for a timestamp already imported and skips an
item already recorded in it.

Also add an Orders tab to the item detail page, sellable items only -
every PurchaseOrderLineItem referencing the item, joined to its PO for
customer and PO number, sorted by order date descending.

+++ src/models.py
@@ -216,7 +216,9 @@ class InventoryCountItem(Base, TimestampMixin):
         ForeignKey("inventory_counts.inventory_count_id"), nullable=False
     )
     item_id: Mapped[int] = mapped_column(ForeignKey("items.item_id"), nullable=False)
-    counted: Mapped[int] = mapped_column(Integer, nullable=False)
+    # A signed quantity, not necessarily whole - some counted items are
+    # weight-based (e.g. 61.80342906 lbs).
+    counted: Mapped[Decimal] = mapped_column(Numeric(14, 8), nullable=False)
     notes: Mapped[str | None] = mapped_column(Text)
 
     inventory_count: Mapped["InventoryCount"] = relationship(back_populates="items")
